omit SPHINCS+ vectors
We are not currently using them and they take 200MB (!) uncompressed,
making the package too large to be uploaded to PyPI and its 100MB limit.
There are optimizations to be made, namely using `bytes` instead of
`strings` when serializing the vectors, and maybe compressing the
resulting protobufs, but for now we will just omit them.