add: test harness mode
Add a new feature to test shared libraries, similar to wrappers. The
shared libraries expose functions with specific names and signatures,
which are described in the documentation. When CC loads the .so and
finds these signatures, it calls the test_lib function of the
corresponding primitive module. This function dlopens the lib, running
the functions with CFFI.

Note that the module and function are called dynamically, skipping the
need to manually add imports. This means that support for this mode can
be added to current and new primitives without having to change the code
outside of the primitive module.

Currently supports AES, SHA, SHAKE, Kyber, and Dilithium.

This commit also updates the usage of the implementations of AES, Kyber,
and Dilithium: instead of compiling an executable, compile them as
shared libraries that get installed in the app data directory and use
them with CFFI. This improves the performance and makes it easier to
incorporate these implementations to the unit tests for the harnesses.

For AES, the segment_size argument is removed. This should be inferred
from the mode of operation used, and removing it allows us to remove an
additional case to support.

We now mark explicitely the version of Kyber and Dilithium used, which
indicates which test vectors are available. A warning is included in the
corresponding documentation pages.