common: save debug data
Add the option to save all the debug data when saving the results to a
file. For now there is no granularity, we either save everything or not.
The idea would be to add an option when tests fail to only include
those. Also maybe add a CLI option, as currently the user is prompted
when accepting to save the results.