simplify the usage of the AES and SHA wrappers
As with ECDH, the user must give the name of the wrapper instead of
using a hardcoded one. The language of the wrapper is inferred from the
file extension (or lack thereof).

The AES C wrapper explicitly malloc's the output array inside the
section that the user must complete, as the size may differ when using
padding for example.

The Makefile is no longer provided, the user must compile the C wrapper
on their own. This simplifies the logic, and allows the user to take
into account other cases, such as linking libraries.