fix ECDSA uncompressed pubkeys
As described in #5, crypto-condor passes (hex) strings to the ECDSA
implementation when using PubKeyEncoding.UNCOMPRESSED instead of bytes,
which is not the intended behaviour. This also causes a crash when
saving the results with debug data, as the __str__ method of SigData
tries to convert the key to a hex string with bytes' hex() method.

To solve this, the hex string is converted to bytes, to match the
expected type. A case is added to test_ECDSA.py::test_verify to confirm
that the correct type is provided to the implementation.

Additionally, test_verify is now aware of whether NIST or Wycheproof
vectors should have been used for testing depending on the elliptic
curve and hash function.

Closes #5