fix(common): pass None as filename to process_results
Console.process_results has a filename argument, that according to its
documentation, if its None the results are only displayed but not saved
to a file, and the user is not prompted for a filename.

As reported in #7, the user is still prompted when passing None. This
commit fixes this, and adds a test to confirm the fix and prevent
regressions.

Co-authored-by: Baptistin BOILOT <baptistin.boilot@ledger.fr>