fix(ECDH): wrong test attribute in Wycheproof test
For context, testing an ECDH implementation with Wycheproof vectors led
to an AttributeError on line 420 that used test.secret instead of
test.private.

Fixing this typo uncovered two additional bugs in
test_exchange_wycheproof regarding the correct interpretation of the
test result.

This showed a problem with the test that should cover that function.
The example wrapper used P-192, but Wycheproof does not provide vectors
for that curve. And a bug in run_wrapper_python meant that Wycheproof
vectors were disabled anyway.

As such, the wrapper example now uses P-256 and the bug is fixed, so the
function is properly covered and verifies that the fixes applied are
working.

Co-authored-by: Baptistin BOILOT <baptistin.boilot@ledger.fr>
Issue: #11