add(cli): --debug option to save debug data
The --save-to option allows users to select a filename for saving the
test results. However, they are still prompted on whether to add the
debug data to this report. This breaks non-interactivity, which is
detrimental to scripting.

We add the --debug and --no-debug options to choose whether or not to
save debug data from the CLI, bypassing the interactive prompt.

Issue: #10