fix(SHA): ensure key uniqueness for test results
Hard-coding the keys in the returned ResultsDict led to a conflict when
testing more than one set of parameters (e.g. SHA-224 and SHA-256). So
we use Results.new and ResultsDict.add, along with the new extra_values
option, to ensure that keys are unique for a given run. This means that
it still does not support combining the results of testing two different
implementations of the same algorithm.

Issue: #8