add(common): new_from_test method for TestInfo
We are currently testing an harmonisation of the test vectors, in which
all tests regardless of primitive have a common set of values such as an
ID and test type. This new method takes the test vector directly to
create a new TestInfo instance, instead of having to unpack its values
into the new() method.