Fix missing test vectors in wheel
Poetry 2.0 [1] has made "include behavior more consistent" [2], meaning
that by default, the include directive only applies for sdist packages.
Since the protobuf test vectors are excluded with .gitignore, they have
to be manually included. And as poetry's version is not pinned in the
CI, when building version 2025.01.09, the test vectors were not included
in the wheel.

Updating pyproject.toml and forcing both formats to include the test
vectors should fix. Tagging release candidate to test the fix.

[1]: https://github.com/python-poetry/poetry/issues/9136
[2]: https://github.com/python-poetry/poetry/issues/9691