Fix unused harness test
pytest searches for files with of the form test_*.py or *_test.py,
meaning that test.py was not included in the test collection and thus
the harnesses were not being tested.

This commit fixes the issue by renaming the test file. Running the test
on the harnesses reveals a bug in the results returned by AES.test_lib,
to be corrected on a follow-up commit.