Update AES test vectors
Remove the vectors/AES.py module to simplify the structure. As with
other primitives, the enums are directly defined in the
primitives/AES.py module.

Update the protobuf definition to generalise the AES test vectors, so
that they can be used for any source and mode of operation. Update the
corresponding classes with protoc.

Update how CAVP files are parsed to use the new AesVectors. Also parse
the Wycheproof files using the same vectors.

Add the JSON file that defines the available test vectors for each
(mode, key length) combination supported.