Update the AES module
Move the enums for vectors.aes.

Update the debug data classes to simplify usage.

Update how the test vectors are loaded, using the JSON file and the new
AesVectors class.

Deprecate the test() function in favour of test_encrypt() and
test_decrypt(). The motivation is twofold: first, as with recent primitives,
offer one function to test one operation instead of trying to have one
function that dispatches tests. This avoids adding too many arguments
and potentially unintuitive None parameters when calling the function.
Second, the internal functions (_test_*_encrypt, _test_*_decrypt) had to
be updated to make use of the new test vectors. Since these vectors do
not differentiate between sources or modes, they can be handled more
easily by a single auxiliary function (_try_one_enc and _try_one_dec).

Deprecate verify_file(), replacing it with test_output_encrypt() and
test_output_decrypt(). Same logic as above, offer one function to test
one operation. It also improves the parsing of the line, matching for
expected patterns and checking whether they correspond to the given
mode.

Mark the handling of invalid tests with FIXME tags. Currently, we only
check if the implementation "failed" when using invalid tests, for a
very large definition of "failed". Ideally, we should use (and maybe
add) the flags to check for specific errors. This could be accomplished
in the Python wrappers with custom Exceptions.

Update the Python runner to use a new naming convention, similar to the
harness API, making use of the new test_encrypt() and test_decrypt()
functions.

Update the internal harness function to call the correct tests.