Update the Python AES wrapper
Add the new calling convention, allowing to test more than one set of
parameters with a single wrapper.

Create a new section in the docs to make it easily discoverable, as the
wrappers are not strictly linked to the Python API, as they can be used
through the CLI.

Update the corresponding command, removing the parameters that are now
encoded in the names of the functions.