Update the SHAKE module
First, update the test vectors. Move the source files from _sha to
_shake, add a protobuf descriptor and the ShakeTest and ShakeVectors
classes, and create a parsing script that also generates the JSON
descriptor.

The protobuf uses the new format with some differences from the one
included in the SHA module: we combine the tests from ShortMsg, LongMsg,
and VariableOut files into the `tests` field, and add the `mc_test`
field containing the Monte-Carlo test. This way, all tests for a given
algorithm and orientation fit in one single instance of ShakeVectors.
Note that users of these vectors (us) are expected to check whether this
is a valid test before using it, as sources other than NIST CAVP are not
expected to provide this one.

Add the `test_digest` function to replace `test`, following the idea of
having one function to test one operation. The function should also
explicitly indicate the operation being tested. The function itself is
updated to make use of the new ShakeVectors. It now returns a single
Results instance for a given (algorithm, orientation) combination.

Add the `test_output_digest` function, which adds support for the output
mode.

Update the Python wrapper using the naming convention of the harness.
Add the wrapper-api section to make the interfaces discoverable, instead
of having to search for them in the Python API.

Add the SHAKE method guide.