Add SHAKE vectors
Add a separate module for SHAKE vectors, including a protobuf
descriptor and classes, a parsing script, and the source files.

The protobuf uses the new format with some differences from the one
included in the SHA module: we combine the tests from ShortMsg, LongMsg,
and VariableOut files into the `tests` field, and add the `mc_test`
field containing the Monte-Carlo test. This way, all tests for a given
algorithm and orientation fit in one single instance of ShakeVectors.

Note that users of these vectors (us) are expected to check whether this
is a valid test before using it, as sources other than NIST CAVP are not
expected to provide this one.