Update SHA test vectors
Remove the bit-oriented test vectors. The current protocol only passes
the message to hash without indication of its length in bits, making it
impossible to test bit-oriented implementations. Furthermore, it seems
that not many implementations are bit-oriented [1], which is anecdotally
confirmed by the absence of bug reports on this mode.

In a first time, the main user-visible change is the fact that `test`
raises ValueError if the given orientation is BIT. The next step is to
remove the orientation completely. If there are users of this mode, we
may add them back. In that case, we should create separate test vectors
with an additional `len` field, as well as a separate test function.

For users of the test vectors, the ShaVectors class from `vectors.SHA`
is replaced by the protobuf class `vectors._sha.sha_pb2.ShaVectors`. The
test vectors themselves are simplified, as we did with the SHAKE
vectors. The tests from the LongMsg and ShortMsg files are combined into
a single instance of ShaVectors, along with a new mc_test field for
Monte Carlo tests. This means that only a single Results is returned per
test vectors file.

[1]: https://csrc.nist.rip/groups/STM/cavp/documents/shs/shaval.html