Update ECDH test vectors.
Update the protobuf descriptor to match the new format. As part of this
update, we change how the tests are stored: the private value is still
stored as bytes, but the public value is now one of two bytes fields,
peer_point and peer_x509. The first is an encoded point representing the
public key, while the second one is a X.509 SubjectPublicKeyInfo. This
corresponds to both types of keys provided by Wycheproof. The idea is to
easily use these tests, as well as simplify the usage. Constructing a
key from the (x, y) coordinates is possible with both cryptography and
PyCryptodome, but less simple with e.g. OpenSSL. Using encoded points
should be more portable than the raw coordinates (see [1]).

The ecpoint files from Wycheproof are also included, since we now have
proper support for public keys as encoded points.

The next step is then to update the protocols to stop using the
coordinates and only provide the encoded point. We could even convert
the point to an X.509 key -- which is the main reason why `parse_cavp`
encodes the point with `ec` instead of manually crafting it. This would
allow us to have compliance vectors for both types of key.

[1]: https://github.com/openssl/openssl/issues/16270#issuecomment-895734092