Update the ECDH module
Move `Curve` to vectors.ecdh. Keeping these enums separately from the
primitive module allow us to use them in e.g. the _import scripts
without encountering circular dependency problems. It may also allow us
to provide lighter enum-based auto-completion in the CLI, by not having
to import the primitive module with all its dependencies.

Remove `nist` and `wycheproof` attributes from `Curve`. Since we are now
using generic vectors, no need to keep these attributes. If needed, we
can query the `source` field of `EcdhVectors`.

Remove the `ec_curve` property. A lighter way to provide this
functionality is simply to return the *name* of the curve as defined in
`ec`. Not only it simplifies the function (not having a `case` for each
and every curve), the definition of `Curve` does not depend on `ec`,
removing a dependency for modules that do not use this method.

Return `ec` curve name for all curves. Currently this method is used for
curves tested by NIST vectors to convert the coordinates to an encoded
point. We can return the name for all *supported* curves for no
additional cost.

Remove attrs-based `EcdhVectors`, as well as `EcdhNistVectors`. We now
define a protobuf-based class `EcdhVectors` that can be used by any
source, so there is no need for separate protobuf classes or a wrapper
with a `load` method. Instead, vectors are loaded with `_load_vectors`.

Replace `EcdhNistData` and `EcdhWycheproofData` with `PointData` and
`X509`. New data classes adapted to the new test vectors.

Update `exchange_nist` and `exchange_wycheproof` to use the new test
vectors. Self-explanatory, but we also now return a `ResultsDict`
instead of `Results | None` as there may be more than one test vectors
file to use.