Update ECDH wrapper
The Python wrapper uses the new naming convention, defined under
wrapper-api, and it uses the new functions introduced. Since the new
test_wrapper functions do not have a `curve` parameter, we can no longer
test the `CC_ECDH` class, at least from the CLI.

As explained in the note in `test_wrapper_python`, we still look for the
class but instead of testing it, we warn about its removal from the
wrapper. To maintain some compatibility while the ECDH protocol exists,
we provide an explanation on how to use the Python API to test the
class, which works for now thanks to `test_exchange`.

The wrapper example is updated to test these new protocols, the second
example is removed as it is not needed.