Update HMAC module
On top of updating the test vectors and the functions that use them, we
introduce two new protocols: `Digest` and `Verify`. These replace `HMAC`
and `HMAC_IUF` with the intention to simplify the API and continue the
idea of 1 operation = 1 test. The former functions and protocols are
deprecated but kept for compatibility.

To use the new protocols, we add the `test_digest` and `test_verify`
functions. These also are designed to use the test vectors in a generic
way, no longer depending on their source.

To match this, the wrapper API is updated to be harness-like and
described in the documentation.