Update AES harness
While the internal implementation does use a single buffer for the
plaintext and ciphertext, it is likely more common that implementations
use two separate buffers. This is particularly the case when padding is
involved, as the buffers will have different sizes. So we update the
function signature to have them separately.

This commit updates documentation page to detail encryption and
decryption, authenticated or not, separately. It adds a test harness for
verifying that the feature works, and each harness is included in its
corresponding section.