Fix TestU01 results parsing when warning occurs
The current parsing relied on skipping a number of lines, assuming that
results are always preceded by the same number of lines. This is not the
case when a warning appears, such as when the file contains mainly
zeroes.

The parsing is improved by looking for a line that appears just before
any results. This means that warnings are ignored for now, but should be
parsed in a future update.

Checking whether the test passed or not is also improved: instead of
relying on the position of the word 'PASS', we check if it's present in
the line.

Issue: #21