Update AES decrypt harness
The harness function returns the length of the decrypted plaintext,
using -1 as an error code. This value is used to truncate the returned
buffer before returning it to test_decrypt(). The corresponding function
is added to the harness test, and resilience test vectors are now also
used when calling test_harness().

The decryption harness allocates a buffer for the plaintext that is the
same size as the ciphertext. For CBC-PKCS7, this is longer than required
due to the padding that should be stripped by the implementation. Thus,
comparing the entire buffer to the reference plaintext would fail, as
the buffer ends with bytes that the implementation may have not written.

Following OpenSSL's example, we use the return value to indicate the
size of the decrypted plaintext, with -1 as the new error code.

Activating resilience test vectors is necessary to actually test the
harness, as CBC-PKCS7 only has Wycheproof test vectors now. This in turn
causes the GCM harness to fail, as Wycheproof has one (1) test vector
with an IV that is longer than 128 bytes, which is rejected by OpenSSL.

For now mark that test with xfail, knowing that all other test values
pass, but it may be worth at least separating the CCM harness from the
GCM one.