Add list_functions to harness
The current way of listing the functions that are recognized by
`test_harness` is by setting the logger's level to DEBUG (INFO does not
cover included functions). This is especially not given to the Python
API users, as increasing the verbosity is mainly a CLI feature. Since
there are many reasons why CC does not correctly detect a harness
function (such as typos or C++ name mangling), we add a way of listing
all functions found in a harness.

First at the API level by introducing `list_functions`, which takes the
function parsing logic from `test_harness` to avoid code duplication.
The output is structured for both `test_harness`, by separating each CC
function found in a per-primitive list, and other users by separating
all functions into three categories: included, excluded, and other.

Then, the CLI command gains a --list option that only calls
`list_functions` and formats the result into a table for each category.
The table box style was chosen to avoid decorations on the sides.

Additionally, the command help message is updated to make it clearer
that all CC functions are tested by default (e.g. when not using
--include or --exclude).