Add HQC
This commit adds support for HQC. This includes both support for Python
wrappers and harnesses.

Some work is required to integrate the reference implementation. As
such, only decapsulation is supported for now, as encapsulation is a
non-deterministic operation that requires an implementation to verify
its result. Nevertheless, we can to test both operations together:
encapsulating then decapsulating with the same implementation should
return the same shared secret. So we introduce the `test_invariant`
function that uses both operations with key pairs taken from compliance
test vectors.

For the purpose of testing the invariant, both wrapper and harness
define an encapsulation function, along with the corresponding
decapsulation. A stub function is added to indicate crypto-condor to
test this invariant.

The test vectors are generated with the reference implementation, using
the included script for NIST KATs.

The index page for PQC algorithms is updated to reflect the selection of
HQC as 5th standard.