Add error message when no test vectors are loaded
Show a message when there are no test vectors for the given parameters.

Depending on the parameters for the primitive and the sources
(compliance/resilience), there may not be any test vectors available.
This is partially checked by the test function itself, as
_load_vectors only filters by parameters and not by source. To simplify
detecting when no test vectors can actually be used, filtering by source
is moved to _load_vectors. This way, the test function can check if the
returned list is empty and display an error message.

Primitives that do not have a _load_vectors function yet are not
updated, but the same logic should be added once the function is added.

The template is also updated.