Improve logger calls
Since currently the default level is WARNING, upgrade some info() calls
to warning().

For try/catch statements in parts such as compiling shared libs use
exception() instead of error() + debug(exc_info=True) to always show the
user the traceback, in order to debug the issue.

Downgrade some critical() logs to exception() to 1) add the traceback,
since they are in try/except statements, and 2) the docs state that
CRITICAL level is used for errors that may stop the program. If CC can't
compile these C libs, some functions _are_ unavailable, but in general
we may want to continue, since a primitive module can't tell if there
are other tests pending (such as testing multiple primitives in a
harness).