Add SLH-DSA
Add support for SLH-DSA, both the pure and pre-hash variants. Signing in
SLH-DSA has two versions: hedged (randomized) and deterministic. We only
support testing deterministic signing for now, unless the signing
operation is tested along the verification with test_invariant().

This is because deterministic signing can be tested directly using test
vectors, whereas hedged signing requires a reference implementation to
actually verify the signatures produced.

The naming convention (and thus the parsing logic) is more convoluted
than usual, on account of the number of choices available for SLH-DSA,
namely:

- the hash function or XOF (SHA-256 or SHAKE-256);
- the security strenght (128, 192, or 256);
- the signature "preference" (short or fast signatures);
- the variant (pure or pre-hash);
- the signing variant (hedged or deterministic).

Tests vectors come from the ACVP server [1], which provides examples of
SLH-DSA test vectors for the three operations. These are parsed and
stored in protobufs as per usual.

The SPHINCS+ guide becomes the SLH-DSA guide, with updated references,
overview, and parameters that match the FIPS specification.

Both wrapper and harness mode are supported. Output mode is not yet
supported, as the module lacks the integration of a reference
implementation.

There are tests for the harness, in all the configurations
(pure/pre-hash, all parameter sets, including the invariant) but it is
*not* yet part of the CI run. There are two reasons:

1) These tests use OpenSSL's implementation of SLH-DSA, which was added
   in version 3.5, in early April 2025. This version is not yet shipped
   in stable versions of distros (afaik), let alone available in the CI
   environement. Instead, it has been compiled and tested manually. The
   plan is to do so in the CI.
2) Signing, especially with all parameter sets, is quite slow. Manually
   testing both pure and pre-hash harnesses takes about 6 minutes, which
   is roughly the average time the rest of the test suite takes to run
   in the CI. A suitable testing strategy is yet to be defined.

[1]: https://github.com/usnistgov/ACVP-Server