Add output mode for ECDH
To simplify the code, this mode only deals with DER keys, which we
expect to be more common than public key coordinates.

To perform the exchange, CC imports the public key using cryptography's
ec module, which does not support all the curves defined by ecdh.Curve.
As such, the CLI has its own enum to define only the curves that can
actually be tested.

Since the public keys must have a specific format, we use values from
the test vectors instead of generating randomly, which would increase
the execution time of the tests.