Fix usage of compliance/resilience options in some primitives
The compliance and resilience options were added to `test_lib` but were
not being passed to the internal _test functions for ECDH, HMAC, and
SHAKE, despite the fact that some of these primitives have resilience
test vectors.

Other primitives use these values correctly, except the post-quantum
ones. We do not currently have resilience test vectors, so we'll update
in a separate commit.