Fix typer conflict with click 8.2.1
Typer 0.9.4 is not compatible with click 8.2.1 due to a missing
positional argument in rich_utils.py:370, so update Typer to the latest
version.

This introduced a breaking change, as the rich_utils module imported to
override the dim text in help messages is now lazy-loaded [1]. The
solution is to directly import rich_utils from typer.

Typer was not updated due to using the caret requirement [2]. The
constraint stays for now, considering potential breaking changes
introduced in new versions. However, the constraint on lief changes
~=0.16, a compatible release [3] allowing >=0.16.0 and <1.0.0.

[1]: https://github.com/fastapi/typer/pull/1128
[2]: https://python-poetry.org/docs/dependency-specification/#caret-requirements