Add partial support for Ed25519
Add tests for signing and verifying functions for Ed25519. Also add
support for testing the output of a signing function, as well as Python
harness support. Tests for these features are included and are passing.
The Python harness comes with a template and example.

Missing from this commit are support for C harness and a method guide.