Update ChaCha20 module
Separate the Encrypt/Decrypt protocols into two versions, one for
ChaCha20 and the other for ChaPoly. This avoids using overloads and is
simpler to type check. Similarly, separate the test functions for each
operation and variant.

Update the harness functions to the current API. Add tests with OpenSSL
for both variants and operations.

For the CLI, update the `test wrapper` command to take the path of the
wrapper, and remove unused options. Remove the --no-save option: use the
filename argument to determine whether to save or not without prompting
the user. For the `test output` command, call the underlying API
directly.