Add test_raw and test_generator to TestU01
test_raw() works as a simple wrapper over test_file(), saving the output
to a temporary file.

test_generator() takes in a function that generates a certain amount of
bytes. The function is called until the desired amount has been
generated and then written to a file to test with test_file().

Add information on how much data to generate, settle on 100KB as the
strict minimum that TestU01 should accept, with 10MB the recommended
minimum.

Deprecate the bit_count argument of test_file(). It checks that the size
of the file is at least 100KB and raises ValueError if not, so bit_count
is no longer used. Will be removed later.