Remove prefix from flag notes names
The Results class adds a prefix ("Compliance" or "Resilience") to flags
when adding a result, to allow the users to differentiate the type of
tests that work or not. Flags are described using notes, which are
matched by the flag name. This means the notes must also have the
prefix, which is not the case for all test vectors.

The solution implemented here is to remove the prefix when saving a set
of all flags used in the current file. This way they are simply matched
to the notes, while the count of tests (valid, invalid, acceptable)
still gets the full flag.