Add test_keygen to ECDSA
Small-ish update of the key generation test for ECDSA to match the new
(minimum) limits for TestU01. It also adds a documentation note to
explain that randomness testing is best-effort, since EC keys are _not_
drawn directly from a (CS)PRNG. The previous test_key_pair_gen() is not
outright replaced but deprecated first.