Add CommonHash to common.py
CommonHash is an enum without members, to provide common methods and
properties for hash functions used in CC. Since it has no members, it
can be inherited by other classes, allowing each primitive to defined
its own allowed hash functions.

Add common.py to the documentation, and include the TESTU01_* constants.
Since we want to display their values, remove `no-value` from the global
conf and add it where necessary (SUPPORTED_MODES).

Following the test with TestU01.py, the documentation can be included in
the module docstring and displayed using the `automodule` directive. The
actual documentation page only requires a title and this directive.