Add test_harness_python to RSAES
Add harness API for Python (did not have a C wrapper before). Combine
the wrapper examples into one file with both types of padding.

Still uses old type of test vectors, importing directly from
Wycheproof's JSON files. This is due to how test vectors are parsed
without keeping the `testGroups`: for most primitives this simplifies
the structure we use, but for RSA it also means storing larger keys for
each test. A possible solution is to create a protobuf file per group,
adding the key to the parameters.